deploy/vertex_predictor/requirements.txt
Flask>=3.0.0
google-cloud-storage>=2.18.2
joblib==1.5.3
numpy==2.4.4
pandas==3.0.2
scikit-learn==1.8.0
